requirements.txt
google-api-python-client
google-auth-oauthlib
wipac-keycloak-rest-services
unidecode
colorlog
